Tweak: Extracted Stretch Section handler to more generic Stretch Element
