+++ assets/dev/js/frontend/modules.js
@@ -1,6 +1,7 @@
 import elementorModules from '../modules/modules';
 import Document from './document';
 import StretchElement from './tools/stretch-element';
+import StretchedElement from './handlers/stretched-element';
 import BaseHandler from './handlers/base';
 import SwiperBase from './handlers/base-swiper';
 import NestedTabs from 'elementor/modules/nested-tabs/assets/js/frontend/handlers/nested-tabs';
@@ -12,6 +13,7 @@ elementorModules.frontend = {
 	},
 	handlers: {
 		Base: BaseHandler,
+		StretchedElement,
 		SwiperBase,
 		NestedTabs,
 	},
